+++ lib/superset_bridge.dart
@@ -1,4 +1,5 @@
 library superset_bridge;
 
+export 'src/superset_bridge_config.dart';
 export 'src/superset_bridge_controller.dart';
 export 'src/superset_bridge_html_content.dart';